requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# ML and data processing
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# Embeddings
sentence-transformers==2.2.2
torch>=2.0.0

# OpenAI for GPT-4V
openai==1.12.0

# Airtable
pyairtable==2.2.0

# Image processing
Pillow==10.2.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
httpx==0.26.0

# Testing
pytest==8.1.1
